fix: moved hard coded colors to Color.kt
